% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataproc.R
\name{eq_clean_data}
\alias{eq_clean_data}
\title{Read raw NOAA data of earthquake and returns a clean data frame
The clean data frame contains:
1. A date column created by uniting the year, month, day and converting it to the Date class
2. LATITUDE and LONGITUDE columns converted to numeric class
3. LOCATION_NAME column by stripping out the country name (including the colon) and converts names to title case (as opposed to all caps).
The  function eq_location_clean() is integrated into eq_clean_data() function to aviod
redundent functions.}
\usage{
eq_clean_data(raw_data_fn = file.path(system.file("extdata", package =
  "MasterRcap"), "signif.txt"))
}
\arguments{
\item{raw_data_fn}{A path to the raw data filename, default to the signif.txt delivered by the
package}
}
\value{
A clean data.frame
}
\description{
Read raw NOAA data of earthquake and returns a clean data frame
The clean data frame contains:
1. A date column created by uniting the year, month, day and converting it to the Date class
2. LATITUDE and LONGITUDE columns converted to numeric class
3. LOCATION_NAME column by stripping out the country name (including the colon) and converts names to title case (as opposed to all caps).
The  function eq_location_clean() is integrated into eq_clean_data() function to aviod
redundent functions.
}
\examples{
\dontrun{
  clean_data <- read_eq_clean_data()
}
}
