% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MapLeaflet.R
\name{eq_create_label}
\alias{eq_create_label}
\title{Creates a label for leaflet map}
\usage{
eq_create_label(data)
}
\arguments{
\item{data}{A data frame containing cleaned NOAA earthquake data}
}
\value{
A character vector with labels
}
\description{
This function creates a label for the \code{eq_map} leaflet map 
based on location, name, magnitude and fatalities from NOAA earthquake data.
}
\details{
The input \code{data.frame} needs to include columns LOCATION_NAME,
EQ_PRIMARY and TOTAL_DEATHS with the earthquake location, magintude and
total casualties respectively.
}
\examples{
\dontrun{
eq_create_label(data)
}
}
