% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataPrep.R
\name{eq_location_clean}
\alias{eq_location_clean}
\title{Cleans earthquake location data}
\usage{
eq_location_clean()
}
\arguments{
\item{data}{A data frame with raw data obtained from NOAA website (see details)}
}
\value{
A data frame with cleaned LOCATION_NAME column
}
\description{
\code{eq_location_clean} transforms NOAA data frame LOCATION_NAME
column by trimming the country name when applicable.
}
\details{
The function requires raw date obtained from NOAA site
\url{https://www.ngdc.noaa.gov/nndc/struts/form?t=101650&s=1&d=1}.
}
\examples{
\dontrun{readr::read_delim(file=file.path("data","results"), delim='\\t') \%>\%
eq_location_clean()}
}
