% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataproc.R
\docType{data}
\name{GeomTimeline}
\alias{GeomTimeline}
\title{Custom Geom ggproto for the timeline}
\format{An object of class \code{GeomTimeline} (inherits from \code{Geom}, \code{ggproto}) of length 5.}
\usage{
GeomTimeline
}
\description{
Custom Geom ggproto for the timeline
}
\details{
This ggproto is for internal use only and not exported to run individually
}
\keyword{datasets}
