% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeomDefn.R
\docType{data}
\name{GeomTimelineLabel}
\alias{GeomTimelineLabel}
\title{Earthquake Timeline Label}
\format{An object of class \code{GeomTimelineLabel} (inherits from \code{Geom}, \code{ggproto}) of length 5.}
\usage{
GeomTimelineLabel
}
\arguments{
\item{n_max}{An integer. If used, it only plots the labels for the
\code{n_max} largest earthquakes in the selected group in the timeline. Also,
the timeline must have size as input parameter.}
}
\description{
This geom plots timeline labels of earthquakes, and assumes that
\code{geom_timeline} was used to create the timelines
}
\details{
This geom function relies on the new Geom class called GeomTimelineLabel
that is created using \code{ggplot2::ggproto}. The required aesthetics for this 
geom is \code{label} that should contain string for labeling each data point.
}
\examples{
   \dontrun{readr::read_delim(file=file.path("data","results"), delim='\\t') \%>\% 
            eq_clean_data() \%>\% 
            dplyr::filter(COUNTRY == c("CHINA","JAPAN"), YEAR >= 2000) \%>\%
            ggplot2::ggplot(ggplot2::aes(x = DATE,
                            y = COUNTRY,
                            color = as.numeric(TOTAL_DEATHS),
                            size = as.numeric(EQ_PRIMARY))) +
              geom_timeline() +
              geom_timeline_label(aes(label = LOCATION_NAME),n_max=3)}
}
\keyword{datasets}
