% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataPrep.R
\name{eq_clean_data}
\alias{eq_clean_data}
\title{Cleans earthquake data from NOAA}
\usage{
eq_clean_data(RawMEdf = MajorEarthquakesAD)
}
\arguments{
\item{data}{A data frame with raw data obtained from NOAA website (see details)}
}
\value{
A data frame with cleaned date, latitude, longitude and location
columns
}
\description{
\code{eq_clean_data} cleans date, latitude and longitude, and location name
from the source NOAA data
}
\details{
The function requires raw date obtained from NOAA site
\url{https://www.ngdc.noaa.gov/nndc/struts/form?t=101650&s=1&d=1}. It adds
a column DATE with cleaned date (Date format), transforms LATITUDE and
LONGITUDE columns as numeric objects and transforms LOCATION_NAME by removing
the country.
}
\examples{
\dontrun{readr::read_delim(file=file.path("data","results"), delim='\\t') \%>\%
eq_clean_data()}
}
