% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeomDefn.R
\docType{data}
\name{GeomTimeline}
\alias{GeomTimeline}
\title{Earthquake Timeline Visualization}
\format{An object of class \code{GeomTimeline} (inherits from \code{Geom}, \code{ggproto}) of length 5.}
\usage{
GeomTimeline
}
\description{
This is a geom function plots earthquake in a timeline. It has the options
to take additional plotting input like country, magnitude of earthquakes and
number of fatalities.
}
\details{
This geom function relies on the new Geom class called GeomTimeline
that is created using \code{ggplot2::ggproto}. This new class specifies a number 
of attributes and functions that describe how data should be drawn on a plot.
}
\examples{
   \dontrun{readr::read_delim(file=file.path("data","results"), delim='\\t') \%>\% 
            eq_clean_data() \%>\% 
            dplyr::filter(COUNTRY == c("CHINA","JAPAN"), YEAR >= 2000) \%>\%
            ggplot2::ggplot(ggplot2::aes(x = DATE,
                            y = COUNTRY,
                            color = as.numeric(TOTAL_DEATHS),
                            size = as.numeric(EQ_PRIMARY))) +
              geom_timeline()}
}
\keyword{datasets}
