% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeomDefn.R
\name{theme_timeline}
\alias{theme_timeline}
\title{Theme for better timeline visualization in ggplot2}
\usage{
theme_timeline()
}
\description{
This theme function provides better rendering for
\code{\link{geom_timeline}}
}
\examples{
   \dontrun{readr::read_delim(file=file.path("data","results"), delim='\\t') \%>\%
            eq_clean_data() \%>\%
            dplyr::filter(COUNTRY == c("CHINA","JAPAN"), YEAR >= 2000) \%>\%
            ggplot2::ggplot(ggplot2::aes(x = DATE,
                            y = COUNTRY,
                            color = as.numeric(TOTAL_DEATHS),
                            size = as.numeric(EQ_PRIMARY))) +
              geom_timeline() +
              geom_timeline_label(aes(label = LOCATION_NAME),n_max=3) +
              theme_timeline() +
              labs(size = "Earthquake Magnitude", color = "Fatalities")}
}
